HW3/Machine Learning hw3/requirement.txt
numpy
Pillow
matplotlib
scikit-learn
